+++ includes/commands/commands.hpp
@@ -1,8 +1,11 @@
 #ifndef COMMANDS_HPP
 # define COMMANDS_HPP
 # include <deque>
+
+# include <vector>
 # include <string>
 # include <iostream>
+# include "channel.hpp"
 # include "user.hpp"
 # include "utils.hpp"
