+++ includes/user.hpp
@@ -1,0 +1,11 @@
+#ifndef USER_HPP
+# define USER_HPP
+
+class User
+{
+	public:
+	User() {};
+	~User() {};
+};
+
+#endif
